third_party/blink/web_tests/plugins/webview-plugin-type-change.html

<!doctype HTML>
<style>
  #plugin {
    width: 150px;
    height: 150px;
    border: 1px solid black;
  }
</style>
<embed id="plugin"></embed>
<script>
document.querySelector('embed').addEventListener('message', function(message) {
    if (message.data == 'loaded')
        if (window.testRunner)
            testRunner.notifyDone();
});
onload = function() {
    document.getElementById('plugin').setAttribute('type', 'application/x-blink-test-plugin');

    if (window.testRunner)
        testRunner.waitUntilDone();
}
</script>
